Initial commit for typescript rewrite
