+++ .github/FUNDING.yml
@@ -1,2 +1,2 @@
 github: KasperiP
-ko_fi: KasperiP
+ko_fi: Kasperi
