+++ .gitignore
@@ -1,5 +1,4 @@
 # Ignore
 node_modules
 .env
-.vscode
 config.json
