+++ .prettierrc.json
@@ -1,8 +1,5 @@
 {
-    "singleQuote": true,
-    "printWidth": 120,
-    "trailingComma": "all",
-    "endOfLine": "auto",
-    "arrowParens": "avoid",
-    "tabWidth": 4
+  "semi": true,
+  "singleQuote": false,
+  "trailingComma": "none"
 }
