+++ .vscode/settings.json
@@ -1,0 +1,23 @@
+{
+  "editor.fontLigatures": true,
+  "[typescriptreact]": {
+    "editor.defaultFormatter": "esbenp.prettier-vscode"
+  },
+  "workbench.tree.renderIndentGuides": "always",
+  "editor.suggestSelection": "first",
+  "references.preferredLocation": "view",
+  "editor.rulers": [80],
+  "editor.formatOnSave": true,
+  "eslint.packageManager": "yarn",
+  "search.exclude": {
+    "**/node_modules": true,
+    "**/build": true,
+    "**/.next": true,
+    "**/yarn.lock": true,
+    "**/coverage": true
+  },
+  "editor.codeActionsOnSave": {
+    "source.fixAll.eslint": true,
+    "source.organizeImports": true
+  }
+}
