--- Dockerfile
@@ -1,13 +1,0 @@
-FROM node:16
-
-# Create the bot's directory
-RUN mkdir -p /usr/src/bot
-WORKDIR /usr/src/bot
-
-COPY package.json /usr/src/bot
-RUN npm install
-
-COPY . /usr/src/bot
-
-# Start the bot.
-CMD ["node", "bot.js"]
