+++ example-config.json
@@ -1,45 +1,45 @@
 {
-    "BOT_TOKEN": "BOT_TOKEN",
-    "MONGODB_URI": "MONGODB_URI",
-    "CLIENT_ID": "CLIENT_ID",
-    "GUILD_ID": "GUILD_ID",
-    "STAFF_ROLES": ["STAFF_ROLE_ID"],
-    "ROLE_CHANNEL": "ROLE_CHANNEL_ID",
-    "COLORS": {
-        "SUCCESS": "#55FF55",
-        "ERROR": "#FF5555",
-        "WARNING": "#FFFF55"
+  "BOT_TOKEN": "BOT_TOKEN",
+  "MONGODB_URI": "MONGODB_URI",
+  "CLIENT_ID": "CLIENT_ID",
+  "GUILD_ID": "GUILD_ID",
+  "STAFF_ROLES": ["STAFF_ROLE_ID"],
+  "ROLE_CHANNEL": "ROLE_CHANNEL_ID",
+  "COLORS": {
+    "SUCCESS": "#55FF55",
+    "ERROR": "#FF5555",
+    "WARNING": "#FFFF55"
+  },
+  "BAN_ROLE": "BAN_ROLE",
+  "BAN_CHANNEL": "BAN_CHANNEL",
+  "MUTE_ROLE": "MUTE_ROLE",
+  "TICKET_CHANNEL": "TICKET_CHANNEL_ID",
+  "TICKET_CATEGORY": "TICKET_CATEGORY_ID",
+  "WARN_EXPIRES": 30,
+  "WARN_THRESHOLD": 4,
+  "WARN_BAN_DAYS": 30,
+  "DELETE_INVITES": true,
+  "SPAN_HANDLER": false,
+  "SELF_ROLES": [
+    {
+      "ID": "ROLE_ID",
+      "NAME": "Minecraft",
+      "EMOJI": "⛏"
     },
-    "BAN_ROLE": "BAN_ROLE",
-    "BAN_CHANNEL": "BAN_CHANNEL",
-    "MUTE_ROLE": "MUTE_ROLE",
-    "TICKET_CHANNEL": "TICKET_CHANNEL_ID",
-    "TICKET_CATEGORY": "TICKET_CATEGORY_ID",
-    "WARN_EXPIRES": 30,
-    "WARN_THRESHOLD": 4,
-    "WARN_BAN_DAYS": 30,
-    "DELETE_INVITES": true,
-    "SPAN_HANDLER": false,
-    "SELF_ROLES": [
-        {
-            "ID": "ROLE_ID",
-            "NAME": "Minecraft",
-            "EMOJI": "⛏"
-        },
-        {
-            "ID": "ROLE_ID",
-            "NAME": "FiveM",
-            "EMOJI": "👮"
-        },
-        {
-            "ID": "ROLE_ID",
-            "NAME": "Yhteisö",
-            "EMOJI": "✨"
-        },
-        {
-            "ID": "ROLE_ID",
-            "NAME": "Viikkokooste",
-            "EMOJI": "👥"
-        }
-    ]
+    {
+      "ID": "ROLE_ID",
+      "NAME": "FiveM",
+      "EMOJI": "👮"
+    },
+    {
+      "ID": "ROLE_ID",
+      "NAME": "Yhteisö",
+      "EMOJI": "✨"
+    },
+    {
+      "ID": "ROLE_ID",
+      "NAME": "Viikkokooste",
+      "EMOJI": "👥"
+    }
+  ]
 }
