+++ package.json
@@ -1,36 +1,38 @@
 {
-    "name": "moticord",
-    "version": "1.0.0",
-    "description": "Discord bot for Finland's biggest Minecraft-server! ",
-    "main": "bot.js",
-    "scripts": {
-        "start": "nodemon bot.js"
-    },
-    "keywords": [
-        "discord",
-        "bot"
-    ],
-    "author": "kassq",
-    "license": "ISC",
-    "devDependencies": {
-        "eslint": "^8.3.0",
-        "eslint-config-prettier": "^8.3.0",
-        "eslint-plugin-import": "^2.25.3",
-        "eslint-plugin-jsdoc": "^37.1.0",
-        "eslint-plugin-prettier": "^4.0.0",
-        "prettier": "^2.5.0"
-    },
-    "engines": {
-        "node": ">=16.6.0",
-        "npm": ">=7.0.0"
-    },
-    "dependencies": {
-        "@discordjs/builders": "^0.10.0",
-        "@discordjs/rest": "^0.2.0-canary.0",
-        "@psibean/discord.js-pagination": "github:KasperiP/discord.js-pagination#discord.js-v13.5.0-support",
-        "discord-api-types": "^0.25.2",
-        "discord.js": "^13.5.0",
-        "mongoose": "^6.1.2",
-        "node-cron": "^3.0.0"
-    }
+  "name": "puskabot",
+  "version": "1.0.0",
+  "description": "Discord bot for Finland's biggest Minecraft-server! ",
+  "main": "bot.ts",
+  "scripts": {
+    "start": "ts-node src/bot.ts",
+    "pretty": "prettier --write \"./**/*.{ts,json}\""
+  },
+  "keywords": [
+    "discord",
+    "bot"
+  ],
+  "author": "kassq",
+  "license": "GPLv3",
+  "engines": {
+    "node": ">=16.6.0",
+    "npm": ">=7.0.0"
+  },
+  "dependencies": {
+    "@discordjs/builders": "^0.13.0",
+    "@discordjs/rest": "^0.4.1",
+    "@psibean/discord.js-pagination": "^4.0.5",
+    "discord-api-types": "^0.31.2",
+    "discord.js": "^13.6.0",
+    "mongoose": "^6.3.1",
+    "node-cron": "^3.0.0"
+  },
+  "devDependencies": {
+    "@types/node": "^17.0.25",
+    "eslint": "^8.3.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "prettier": "^2.5.0",
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.3"
+  }
 }
